Add files via upload
Updated HOUT code to account for the case when dividing by a zero vector. Specifically in the HOPM3 and HOPM4 codes we replaced the dividing by norm(v1) with dividing by max(eps,norm(v1)).  That way if v1 is the zero vector, it stays the zero vector instead of having a divides by zero problem.